+++ HISTORY.rst
@@ -12,7 +12,10 @@ Future History
 Recent History
 ~~~~~~~~~~~~~~
 
-Next Release
+.. Next Release
+.. ------------------
+
+1.0.0 (2020-11-08)
 ------------------
 
 * Turn on waitlist with manual processing.